+++ main.py
@@ -1,4 +1,4 @@
-from app.window_01 import App
+from app import App
 
 if __name__ == "__main__":
     app = App()